requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
google-adk>=1.0.0
python-dotenv>=1.0
SQLAlchemy>=2.0
pydantic>=1.10
